Backend: Fixed 2 days for small routes bug | Fixed API request data types

+++ hos-eld-trip-planner/hos_eld_trip_planner/trips/serializers.py
@@ -1,6 +1,6 @@
 from rest_framework import serializers
 
 class TripSerializer(serializers.Serializer):
-    dist_btwn_current_and_pickup = serializers.IntegerField(required=True, min_value=0)
-    dist_btwn_pickup_and_dropoff = serializers.IntegerField(required=True, min_value=0)
-    cycles_used = serializers.IntegerField(required=True, min_value=0, max_value=70)
+    dist_btwn_current_and_pickup = serializers.FloatField(required=True, min_value=0)
+    dist_btwn_pickup_and_dropoff = serializers.FloatField(required=True, min_value=0)
+    cycles_used = serializers.FloatField(required=True, min_value=0, max_value=70)
